+++ src/services/facility-service/usecase/vehicle.js
@@ -1,5 +1,7 @@
 const CampusRepository = require('../../../repositories/campus-repository');
 const FacilityRepository = require('../../../repositories/facility-repository');
+const BookingRepository = require('../../../repositories/booking-repository');
+
 const { ImageFacilityStorage } = require('../../../utils/storage');
 const LoggingService = require('../../logging-service');
 const { facilityCategory } = require('../constant');
@@ -91,6 +93,11 @@ class VehicleUsecase {
         return facilityFilter;
     }
 
+    static async __checkFacilityInBooking(id) {
+        const booking = await BookingRepository.getBookingByFacilityId(id);
+        return booking ? true : false;
+    }
+
     static async create(data, files, userId) {
         const facility = await this.__createFacility(
             data,
@@ -164,11 +171,17 @@ class VehicleUsecase {
             };
         }
 
-        const oldData = await this.get(facility.id);
+        if (await this.__checkFacilityInBooking(id)) {
+            return {
+                message: 'Facility Vehicle have booking',
+            };
+        }
 
+        const oldData = await this.get(facility.id);
         const images = vehicle.image || [];
-        await this.__deleteImage(images);
+
         await FacilityRepository.deleteFacility(id);
+        await this.__deleteImage(images);
 
         await catchThrows(
             LoggingService.createLoggingFacility(
